Initial commit; original code by @travis-g &c.
This code was picked up at
<https://codeclimate.com/github/travis-g/Voat-Enhancement-Suite/code>,
where it was "no longer able to connect to [the] repo" and
"out-of-date".  This project was assumed to be abandonware, and this
version of it can be considered either a continuation or a fork.  See
`src/license.js` for all the original licenses and copyright
information.  It's licensed under the WTFPL2.

Signed-off-by: Jonathan Sifuentes <jayands2k11@ymail.com>