NPM and friends
So we have a dependency chart now, using [`yarn`][yarn] and NPM.  The
entries to the `.gitignore` file are for various editors, as well as if
you just need a temp folder to try something in.

[yarn]: https://yarnpkg.com

I know the original used [`grunt`][grunt-cli] (the references to it are
all over the license.js file), but maybe something like Webpack moving
forward? (See #1)

Signed-off-by: Jonathan Sifuentes <jayands@users.noreply.github.com>